Split core_impl into directories for Android, Java 7, and Java 8.
This directory structure will allow us to support the three Java versions but
also take advantage of the features available in each one.  Each directory has a
readme.

I refactored the stats code so that there is one StatsManagerImpl for each of
Android, Java 7, and Java 8.  StatsManagerImpl is loaded with reflection, and it
specifies the stats classes that should be used for the implementation.  I also
moved DisruptorEventQueue to core_impl_java, since it isn't Android-compatible.

This commit doesn't affect tracing.
