+++ core/README.md
@@ -1,0 +1,6 @@
+Instrumentation API
+======================================================
+
+* Java 6 and Android compatible.
+* The abstract classes in this directory can be subclassed to create alternative
+  implementations of the Instrumentation library.
