+++ core/build.gradle
@@ -4,7 +4,6 @@ dependencies {
     compile libraries.grpc_context,
             libraries.guava
     compileOnly libraries.auto_value
-    testCompile libraries.grpc_context
 
     signature "org.codehaus.mojo.signature:java16:+@signature"
 }
