+++ core_impl/README.md
@@ -1,0 +1,5 @@
+Instrumentation implementation
+======================================================
+
+* The main implementation.
+* Java 6 and Android compatible.
