+++ core_impl_android/README.md
@@ -1,0 +1,6 @@
+Instrumentation Android implementation
+======================================================
+
+* Java 6 and Android compatible.
+* StatsManager specifies the stats implementation classes that should be used
+  with Android.
