+++ core_impl_java/README.md
@@ -1,0 +1,5 @@
+Instrumentation Java implementation
+======================================================
+
+* Java 7 compatible.
+* Contains any classes not compatible with Android.
