+++ core_impl_java/src/main/java/com/google/instrumentation/stats/StatsManagerImplJava.java
@@ -11,23 +11,16 @@
  * limitations under the License.
  */
 
-package com.google.instrumentation.common;
+package com.google.instrumentation.stats;
+
+import com.google.instrumentation.common.DisruptorEventQueue;
 
 /**
- * A queue that processes events. See {@code DisruptorEventQueue} for an example.
+ * Java 7 and 8 implementation of {@link StatsManager}.
  */
-public interface EventQueue {
-  void enqueue(Entry entry);
+public abstract class StatsManagerImplJava extends StatsManagerImplBase {
 
-  /**
-   * Base interface to be used for all entries in {@link EventQueue}. For example usage,
-   * see {@code DisruptorEventQueue}.
-   */
-  public interface Entry {
-    /**
-     * Process the event associated with this entry. This will be called for every event in the
-     * associated {@link EventQueue}.
-     */
-    void process();
+  public StatsManagerImplJava() {
+    super(DisruptorEventQueue.getInstance());
   }
 }
