+++ core_impl_java_7/README.md
@@ -1,0 +1,6 @@
+Instrumentation Java 7 implementation
+======================================================
+
+* Java 7 compatible.
+* StatsManager specifies the stats implementation classes that should be used
+  with Java 7.
