+++ core_impl_java_8/README.md
@@ -1,0 +1,5 @@
+Instrumentation Java 8 implementation
+======================================================
+
+* StatsManager specifies the stats implementation classes that should be used
+  with Java 8.
